+++ domain-api/src/main/kotlin/de/chrgroth/spotify/control/domain/outbox/DomainOutboxPartition.kt
@@ -1,12 +1,10 @@
 package de.chrgroth.spotify.control.domain.outbox
 
 import de.chrgroth.spotify.control.util.outbox.OutboxPartition
-import java.time.Duration
 
 sealed interface DomainOutboxPartition : OutboxPartition {
     data object ToSpotify : DomainOutboxPartition {
         override val key = "to-spotify"
-        override val throttleInterval: Duration = Duration.ofSeconds(1)
     }
 
     data object ToSpotifyPlayback : DomainOutboxPartition {