+++ domain-api/src/main/kotlin/de/chrgroth/spotify/control/domain/port/out/PlaybackStatePort.kt
@@ -1,0 +1,5 @@
+package de.chrgroth.spotify.control.domain.port.out
+
+interface PlaybackStatePort {
+    fun onPlaybackDetected()
+}